refactor(M1-I3): 統合前の衝突面を整理
M1-I2 と cases CMS の後続統合で不要な衝突を減らすため、挙動を変えずに型・テスト配置・route登録順を整理する。

Refs: #220
